+++ tests/webhook.js
@@ -28,13 +28,13 @@ suite('Webhook', function () {
 
   });
 
-  test('has the get method', function () {
-    assert.isFunction(webhook.get, 'get method exists');
-  });
+  // test('has the get method', function () {
+  //   assert.isFunction(webhook.get, 'get method exists');
+  // });
 
-  test('has the put method', function () {
-    assert.isFunction(webhook.put, 'put method exists');
-  });
+  // test('has the put method', function () {
+  //   assert.isFunction(webhook.put, 'put method exists');
+  // });
 
   test('has the post method', function () {
     assert.isFunction(webhook.post, 'post method exists');
@@ -53,22 +53,22 @@ suite('Webhook', function () {
     assert.equal(w2.url, '/api/webhook/v0/a_bc/catch');
   });
 
-  test('that get method will send a get request', function (done) {
+  // test('that get method will send a get request', function (done) {
 
 
-    webhook.get().then(function (response) {
-      done();
-    });
+  //   webhook.get().then(function (response) {
+  //     done();
+  //   });
 
-    assert.equal(this.request.method, 'GET');
-    assert.equal(this.request.requestHeaders['Content-Type'], "application/json");
-    assert.equal(this.request.url, '/api/webhook/v0/mywebhook/catch');
+  //   assert.equal(this.request.method, 'GET');
+  //   assert.equal(this.request.requestHeaders['Content-Type'], "application/json");
+  //   assert.equal(this.request.url, '/api/webhook/v0/mywebhook/catch');
 
-    this.request.respond(200, {
-      "Content-Type": "application/json"
-    }, JSON.stringify(response));
+  //   this.request.respond(200, {
+  //     "Content-Type": "application/json"
+  //   }, JSON.stringify(response));
 
-  });
+  // });
 
   test('that get method will send a post request', function (done) {
     webhook.post().then(function (response) {
@@ -85,19 +85,19 @@ suite('Webhook', function () {
 
   });
 
-  test('that get method will send a put request', function (done) {
-    webhook.put().then(function (response) {
-      done();
-    });
+  // test('that get method will send a put request', function (done) {
+  //   webhook.put().then(function (response) {
+  //     done();
+  //   });
 
-    assert.equal(this.request.method, 'PUT');
-    assert.equal(this.request.requestHeaders['Content-Type'], "application/json;charset=utf-8");
-    assert.equal(this.request.url, '/api/webhook/v0/mywebhook/catch');
+  //   assert.equal(this.request.method, 'PUT');
+  //   assert.equal(this.request.requestHeaders['Content-Type'], "application/json;charset=utf-8");
+  //   assert.equal(this.request.url, '/api/webhook/v0/mywebhook/catch');
 
-    this.request.respond(200, {
-      "Content-Type": "application/json"
-    }, JSON.stringify(response));
+  //   this.request.respond(200, {
+  //     "Content-Type": "application/json"
+  //   }, JSON.stringify(response));
 
-  });
+  // });
 
 });